requirements.txt
amqp==5.0.6
ansible==3.3.0
ansible-base==2.10.9
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
Django==3.2.1
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-filter==2.4.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
IPy==1.1
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
packaging==20.9
prompt-toolkit==3.0.18
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
